Move sweepers in separate file
